Bug 931 - ep-engine shutdown may not include some items
Improved the shutdown speed by replacing the 5 second sleep with a loop of
1-sec sleep until the next dirty item shuld be written to disk.

The last flush should also set min_data_age to 0 to cause all dirty items
to be written. If you would like to verify that there is no dirty items
left in memory before shutting down you should export the environment
variable EP_VERIFY_SHUTDOWN_FLUSH. If this variable is set, ep-engine
will inspect all items in memory and ensure that it's dirty-bit is cleared.

+++ ep_engine.h
@@ -408,7 +408,7 @@ public:
                     }
                     if (tc->backfillAge != 0) {
                         sprintf(tap, "%s:backfill_age", tc->client.c_str());
-                        add_casted_stat(tap, tc->backfillAge, add_stat, cookie);
+                        add_casted_stat(tap, (size_t)tc->backfillAge, add_stat, cookie);
                     }
                 }
             }
